Trying to optimize the SKU selector
