+++ components/ui/AvatarSize.tsx
@@ -6,7 +6,7 @@
 const selectedVariants: Record<string, string> = {
   // Color variants - only applied when no color as content is passed
   active: "bg-black text-base-100",
-  disabled: "bg-neutral-content text-neutral",
+  disabled: "text-gray-600 border border-gray-100",
   default: "bg-base-100 text-base-content",
 };
 
@@ -18,13 +18,13 @@ interface Props {
 const variants = {
   active: "",
   disabled:
-    `relative after:absolute after:left-0 after:top-1/2 after:h-[1px] after:bg-red-800 after:w-full after:block after:-rotate-45 after:content-[""]`,
+    `relative after:absolute after:left-0 after:top-1/2 after:h-[1px] overflow-hidden after:bg-neutral after:w-full after:block after:-rotate-45 after:content-[""]`,
   default: "border border-neutral hover:border-primary",
 };
 
 function AvatarSize({ content, variant = "default" }: Props) {
   return (
-    <div class="placeholder text-xs">
+    <div class="placeholder text-xs ">
       <div
         class={`rounded-full py-2 px-4 ${selectedVariants[variant]} ${
           variants[variant]
